ingest/requirements.txt
azure-search-documents==11.2.0
python-dotenv==0.19.0
beautifulsoup4==4.10.0